Renamed ExtraFunctionality and moved some functionality.
- Moved localTime and epoch time generation into its own class for easier readability.

- Renamed ExtraFunctionality to TimerFunctionality.
